refactor(log): 统一日志模块搜索表单风格，优化用户ID字段类型
1.  将操作日志模型的user_id字段从Integer改为BigInteger，新增对应数据库迁移文件
2.  重构登录日志和操作日志的搜索表单，改为标准折叠网格布局
3.  替换登录日志查询参数的status字段为BoolField类型
4.  补充日志模块相关的文档和路由配置

+++ backend/app/models/sys/operation_log.py
@@ -3,7 +3,7 @@
 
 from core.models.base import Base
 from sqlalchemy.orm import mapped_column, Mapped
-from sqlalchemy import String, Text, Integer, Float
+from sqlalchemy import String, Text, Integer, BigInteger, Float
 
 
 class SysOperationLog(Base):
@@ -13,7 +13,7 @@ class SysOperationLog(Base):
     """
 
     user_id: Mapped[int] = mapped_column(
-        Integer, nullable=False, index=True, comment="操作人ID"
+        BigInteger, nullable=False, index=True, comment="操作人ID"
     )
     username: Mapped[str] = mapped_column(
         String(50), nullable=False, comment="操作人用户名"
